refactor(structure): Moved Context to separate folder in project
