+++ src/components/Main-window.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useContext } from "react"
 
-import { Context } from "../Context.jsx"
+import { Context } from "../context/Context.jsx"
 import useMusicPlayerState from "../hooks/useMusicPlayerState.js"
 
 import Nav from './Navbar'
